diplay the menu
